+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Make Release Status fields `Ready` and `InUse` optional.
+- Add printer columns for Release `Ready` and `InUse` fields.
 
 ## [3.8.0] - 2020-11-13