+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Add `catalog` field to `apps` in `release`.
 - Make Release Status fields `Ready` and `InUse` optional.
 - Add printer columns for Release `Ready` and `InUse` fields.
 
@@ -25,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Add `Silence` CRD. 
+- Add `Silence` CRD.
 
 ## [3.6.0] - 2020-11-03