+++ CHANGELOG.md
@@ -9,14 +9,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `config-controller.giantswarm.io/version` label.
+- Add `config.giantswarm.io/major-version` annotation.
+
 ## [3.9.0] - 2020-11-24
+
+### Added
+
+- Add `catalog` field to `apps` in `release`.
+- Add printer columns for Release `Ready` and `InUse` fields.
+- Add printer columns for App, Chart `Version`, `Last Deployed` and `Status`.
 
 ### Changed
 
-- Add `catalog` field to `apps` in `release`.
 - Make Release Status fields `Ready` and `InUse` optional.
-- Add printer columns for Release `Ready` and `InUse` fields.
-- Add printer columns for App, Chart `Version`, `Last Deployed` and `Status`.
 
 ## [3.8.0] - 2020-11-13